Added TODO notes for the final pass.
The TODO notes are there to remind me what to do for this last step in the conversion process. It involves inspecting the <project-name>.xdk file for additional details regarding the Cordova plugins that have been added to the project. Those entries in the final config.xml file need to be updated to reflect their original source, otherwise they all appear as sourced from the NPM registry.