Changed to now require specifying <project-name>.xdk file as a command-line parameter.
The previous method was flawed because it assumed the <name> tag was the same name as the <project-name>, which is not a valid assumption. Thus, the script needs to be told, via a command-line option, what the name of that <project-name>.xdk file is. This change also eliminates the need to CD to the project folder before starting the script, since we can derive the location of the project folder from the <project-name>.xdk file that is given as a command-line parameter.