Must use the Test tab to update files, simply switching projects is not good enough.
There is some processing of the intelxdk.config.additions.xml file that only takes place when you either submit your app to the Test tab or the **old** Build tab (when the XDK build system was still operational). Simply switching projects is not sufficient, all that does is save all files. You have to push the green button on the Test tab to get this processing to take place. Once you do that, all changes to your intelxdk.config.additions.xml file will be processed into the various intelxdk.config.<platform>.xml files, which are the files that the script processes.